backend/src/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
install==1.3.5
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
